backend/pyproject.toml
[project]
name = "cookbooks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dnspython>=2.7.0",
    "fastapi>=0.115.4",
    "flask-bcrypt>=1.0.1",
    "flask-cors>=5.0.0",
    "flask-jwt-extended>=4.6.0",
    "flask-pymongo>=2.3.0",
    "flask>=3.0.3",
    "openai>=1.52.2",
    "pillow>=11.0.0",
    "pydantic>=2.9.2",
    "pymongo>=4.10.1",
    "python-dotenv>=1.0.1",
    "gunicorn>=23.0.0",
    "argparse>=1.4.0",
]
